+++ front/group_user.form.php
@@ -47,16 +47,7 @@ $group_user = new Group_User();
 
 if (isset($_POST["add"])) {
     $group_user->check(-1, CREATE, $_POST);
-    if ($group_user->add($_POST)) {
-        Event::log(
-            $_POST["groups_id"],
-            "groups",
-            4,
-            "setup",
-            //TRANS: %s is the user login
-            sprintf(__('%s adds a user to a group'), $_SESSION["glpiname"])
-        );
-    }
+    $group_user->add($_POST);
     Html::back();
 }